+++ NEWS.md
@@ -8,6 +8,7 @@
 - Added retrying requests to CTIS, which apparently started to refuse or throttle responses 
 - Added incrementing numbers of historic versions of CTIS created with `querytoupdate = ...`
 - Simplified, accelerated loading NDJSON files, minimising deletion and re-loading
+- Changed deleting / preserving temporary files, see also `options(ctrdata.tempdir = "<user_specified_folder>")`
 
 ## Bug fixes
 - Refined indicator variables for `f.hasResults()`